Changing author/contributor designation as per chair guidelines
The group adopts author and contributor definitions as per ICMJE, substantial contributions to the the acquisition, analysis, or interpretation of data for the work; AND drafting the work or revising it critically for important intellectual content; AND final approval of the version to be published; AND agreement to be accountable for all aspects of the work in ensuring that questions related to the accuracy or integrity of any part of the work are appropriately investigated and resolved. see http://www.icmje.org/recommendations/browse/roles-and-responsibilities/defining-the-role-of-authors-and-contributors.html. In that spirit Wolberger is the author. Zundel and Hernandez contributed major sections.